[Refactor] Move profiling methods to MM budget
